backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.0.9
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
python-docx>=1.1.0
pypdf>=3.17.4
tiktoken>=0.5.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
